chore(build): overhaul dev-tasks
Refactor dev-tasks:

* unlock and update dependencies, address changes to them due to those
updates
* improve maintainability/readability
* improve efficiency
* consolidate process-level variables and functions into `context`
object
* miscellaneous fixes
