+++ browser/.jshintrc
@@ -1,0 +1,29 @@
+{
+  "node": true,
+  "browser": false,
+  "globals": {
+  },
+  "esnext": false,
+  "expr": true,
+  "bitwise": true,
+  "boss": true,
+  "camelcase": true,
+  "curly": true,
+  "eqeqeq": true,
+  "forin": false,
+  "freeze": true,
+  "immed": true,
+  "latedef": true,
+  "laxbreak": true,
+  "newcap": true,
+  "noarg": true,
+  "noempty": true,
+  "nonbsp": true,
+  "quotmark": "single",
+  "maxlen": 80,
+  "multistr": true,
+  "regexp": true,
+  "sub": true,
+  "trailing": true,
+  "undef": true
+}
