--- browser/dev-tasks/gulp-marklogic.js
@@ -1,7 +1,0 @@
-/**
- * TODO
- * @type {Object}
- * @return {Object}
- */
-var gml = module.exports = {
-};
