+++ browser/package.json
@@ -16,63 +16,58 @@
     "prepublish": null
   },
   "devDependencies": {
-    "body-parser": "1.5.0",
-    "browserify": "^5.9.1",
-    "chai": "1.9.1",
-    "chai-as-promised": "4.1.1",
-    "chalk": "0.5.1",
-    "connect": "3.1.0",
-    "connect-livereload": "0.4.0",
-    "connect-modrewrite": "0.5.11",
-    "cucumber": "0.4.1",
-    "del": "0.1.1",
-    "duplexer": "0.1.1",
-    "event-stream": "3.1.7",
-    "express": "4.7.2",
-    "globule": "0.2.0",
-    "gulp": "~3.8.7",
-    "gulp-buffer": "0.0.2",
-    "gulp-clone": "1.0.0",
-    "gulp-debug": "0.3.0",
-    "gulp-embedlr": "0.5.2",
-    "gulp-filelog": "0.2.0",
-    "gulp-filter": "0.4.1",
-    "gulp-if": "1.2.4",
-    "gulp-ignore": "1.2.0",
-    "gulp-inject": "0.4.1",
-    "gulp-jscs": "0.6.0",
-    "gulp-jshint": "1.7.1",
-    "gulp-load-plugins": "0.5.3",
-    "gulp-match": "0.2.0",
+    "body-parser": "^1.6.5",
+    "browserify": "^5.10.0",
+    "chai": "^1.9.1",
+    "chai-as-promised": "^4.1.1",
+    "chalk": "^0.5.1",
+    "connect": "^3.1.1",
+    "connect-livereload": "^0.4.0",
+    "connect-modrewrite": "^0.7.7",
+    "cucumber": "^0.4.2",
+    "del": "^0.1.2",
+    "duplexer": "^0.1.1",
+    "event-stream": "^3.1.7",
+    "express": "~4.7.4",
+    "gulp": "^3.8.7",
+    "gulp-clone": "^1.0.0",
+    "gulp-debug": "^1.0.0",
+    "gulp-embedlr": "^0.5.2",
+    "gulp-filelog": "^0.2.0",
+    "gulp-if": "^1.2.4",
+    "gulp-jscs": "^1.1.0",
+    "gulp-jshint": "^1.8.4",
+    "gulp-load-plugins": "^0.5.3",
     "gulp-mocha-phantomjs": "stu-salsbury/gulp-mocha-phantomjs#emit-end",
-    "gulp-plumber": "0.6.4",
-    "gulp-rename": "1.2.0",
-    "gulp-ruby-sass": "0.6.0",
-    "gulp-sass": "0.7.2",
+    "gulp-plumber": "^0.6.4",
+    "gulp-ruby-sass": "^0.7.1",
+    "gulp-sass": "^0.7.3",
     "gulp-tap": "^0.1.1",
-    "gulp-template": "0.1.1",
-    "gulp-util": "3.0.0",
-    "gulp-watch": "0.6.8",
-    "istanbul-middleware": "0.2.0",
-    "jscs-jsdoc": "0.0.6",
-    "jshint-stylish": "0.4.0",
-    "jsonschema": "1.0.0",
-    "lazypipe": "0.2.2",
-    "lodash": "2.4.1",
-    "main-bower-files": "1.0.1",
-    "map-stream": "0.1.0",
-    "marked": "0.3.2",
-    "mocha-phantomjs": "3.3.2",
-    "open": "0.0.5",
-    "phantomjs": "1.9.7-14",
-    "protractor": "1.0.0",
-    "request": "2.40.0",
-    "selenium-webdriver": "2.42.1",
-    "serve-static": "1.4.2",
-    "shelljs": "0.3.0",
-    "through2": "0.4.2",
-    "tiny-lr-fork": "0.0.5",
-    "vinyl-fs": "0.3.4"
+    "gulp-template": "^1.0.0",
+    "gulp-util": "^3.0.0",
+    "gulp-watch": "^0.6.9",
+    "istanbul-middleware": "^0.2.0",
+    "jscs-jsdoc": "^0.0.6",
+    "jshint-stylish": "^0.4.0",
+    "jsonschema": "^1.0.0",
+    "lazypipe": "^0.2.2",
+    "lodash": "^2.4.1",
+    "main-bower-files": "^1.0.3",
+    "map-stream": "^0.1.0",
+    "marked": "^0.3.2",
+    "minimist": "^1.1.0",
+    "mocha-phantomjs": "^3.5.0",
+    "open": "^0.0.5",
+    "phantomjs": "^1.9.7-15",
+    "protractor": "^1.0.0",
+    "request": "^2.40.0",
+    "selenium-webdriver": "^2.42.1",
+    "serve-static": "^1.5.3",
+    "shelljs": "^0.3.0",
+    "through2": "^0.6.1",
+    "tiny-lr-fork": "^0.0.5",
+    "vinyl-buffer": "0.0.0",
+    "vinyl-fs": "^0.3.6"
   },
   "dependencies": {}
 }
